Machinations to avoid exposing async syntax to Py2, because it pukes.
